🎨 Update: refactoring codes
