+++ packages/assignment-7/package.json
@@ -23,7 +23,8 @@
     "framer-motion": "^11.3.2",
     "msw": "^2.3.1",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "zustand": "^4.5.4"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.5",
