+++ packages/assignment-7/src/main.tsx
@@ -5,7 +5,7 @@ import { ChakraProvider } from "@chakra-ui/react";
 
 async function prepare() {
   const { setupWorker } = await import("msw/browser");
-  const { mockApiHandlers } = await import("./mockApiHandlers.ts");
+  const { mockApiHandlers } = await import("./mocks/mockApiHandlers.ts");
   const worker = setupWorker(...mockApiHandlers);
 
   // `worker.start()` returns a Promise that resolves
