+++ packages/assignment-7/src/store/searchTermStore.ts
@@ -1,0 +1,14 @@
+import { create } from "zustand";
+
+type State = {
+  searchTerm: string;
+};
+
+type Action = {
+  setSearchTerm: (term: string) => void;
+};
+
+export const searchTermStore = create<State & Action>((set) => ({
+  searchTerm: "",
+  setSearchTerm: (term) => set({ searchTerm: term }),
+}));
