+++ packages/assignment-7/src/types/index.ts
@@ -1,0 +1,25 @@
+export type RepeatType = "none" | "daily" | "weekly" | "monthly" | "yearly";
+
+export interface RepeatInfo {
+  type: RepeatType;
+  interval: number;
+  endDate?: string;
+}
+
+export type Notification = {
+  id: number;
+  message: string;
+};
+
+export interface Event {
+  id: number;
+  title: string;
+  date: string;
+  startTime: string;
+  endTime: string;
+  description: string;
+  location: string;
+  category: string;
+  repeat: RepeatInfo;
+  notificationTime: number; // 분 단위로 저장
+}
