+++ packages/assignment-7/src/utils/api.ts
@@ -1,0 +1,52 @@
+import { Event } from "../types";
+
+export const fetchEvents = async () => {
+  const response = await fetch("/api/events");
+  if (!response.ok) {
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+  return response;
+};
+
+export const updateEvent = async (eventData: Event) => {
+  const response = await fetch(`/api/events/${eventData.id}`, {
+    method: "PUT",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(eventData),
+  });
+  if (!response.ok) {
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+  return response.json();
+};
+
+export const saveEvent = async (eventData: Event) => {
+  const response = await fetch("/api/events", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(eventData),
+  });
+  if (!response.ok) {
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+  return response.json();
+};
+
+export const deleteEvent = async (id: number) => {
+  try {
+    const response = await fetch(`/api/events/${id}`, {
+      method: "DELETE",
+    });
+
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    return response;
+  } catch (error) {
+    console.log(error);
+  }
+};
